Advanced version of wheel chair
In this project, we transitioned from using the ESP32 to the Raspberry Pi, enhancing our system's capabilities significantly. We also integrated a GPS module to accurately detect the wheelchair's location, enabling real-time tracking and navigation. This upgrade not only improves the reliability of our system but also empowers users with precise location data, ensuring greater safety and independence for individuals using the wheelchair. With these advancements, we aim to create a smarter, more responsive mobility solution that adapts to the needs of its users.